[ADD] major indices data file
