+++ lib/chart/modal_chart.dart
@@ -14,20 +14,16 @@ class KospiModal extends StatelessWidget {
         child: Container(
           height: 300,
           width: MediaQuery.of(context).size.width,
-          //color: Colors.yellow, // 전체 컨테이너에 노란색 배경 적용
           child: Column(
             mainAxisAlignment: MainAxisAlignment.center,
             children: [
               Container(
-                // margin: EdgeInsets.fromLTRB(50, 20, 0, 0),
-                //color: Colors.red, // 텍스트 컨테이너에 빨간색 배경 적용
                 padding: const EdgeInsets.all(10),
                 child: Text(
                   '코스피 차트',
                   style: TextStyle(fontSize: 18, fontWeight: FontWeight.bold),
                 ),
               ),
-              //SizedBox(width: 10), // 빨간색 박스와 파란색 박스 사이에 간격을 줄 수 있습니다.
               Expanded(
                 child: Padding(
                   padding: const EdgeInsets.all(10),
@@ -56,7 +52,6 @@ class KosdaqModal extends StatelessWidget {
           height: 300,
           width: MediaQuery.of(context).size.width,
           child: Column(
-            //crossAxisAlignment: CrossAxisAlignment.start,
             mainAxisAlignment: MainAxisAlignment.center,
             children: [
               Text(
